+++ deno/rest/v9/channel.ts
@@ -382,6 +382,10 @@ export type RESTDeleteAPIChannelMessageUserReactionResult = never;
  */
 export interface RESTGetAPIChannelMessageReactionUsersQuery {
 	/**
+	 * The reaction type
+	 */
+	type?: ReactionType;
+	/**
 	 * Get users after this user ID
 	 */
 	after?: Snowflake;
@@ -391,6 +395,14 @@ export interface RESTGetAPIChannelMessageReactionUsersQuery {
 	 * @default 25
 	 */
 	limit?: number;
+}
+
+/**
+ * https://discord.com/developers/docs/resources/channel#get-reactions-reaction-types
+ */
+export enum ReactionType {
+	Normal,
+	Super,
 }
 
 /**