+++ rest/v9/poll.ts
@@ -20,15 +20,19 @@ export interface RESTGetAPIPollAnswerVotersQuery {
 /**
  * https://discord.com/developers/docs/resources/poll#poll-create-request-object-poll-create-request-object-structure
  */
-export interface RESTAPIPollCreate extends Omit<APIPoll, 'answers' | 'expiry' | 'results'> {
-	/**
-	 * Number of hours the poll should be open for, up to 7 days
-	 */
-	duration: number;
+export interface RESTAPIPollCreate
+	extends Omit<APIPoll, 'allow_multiselect' | 'answers' | 'expiry' | 'layout_type' | 'results'>,
+		Partial<Pick<APIPoll, 'allow_multiselect' | 'layout_type'>> {
 	/**
 	 * Each of the answers available in the poll, up to 10
 	 */
 	answers: Omit<APIPollAnswer, 'answer_id'>[];
+	/**
+	 * Number of hours the poll should be open for, up to 32 days
+	 *
+	 * @default 24
+	 */
+	duration?: number;
 }
 
 /**